conditional noexcept
